requirements.txt
numpy
pandas
matplotlib
opencv-python
OpenEXR
imageio
seaborn
tensorflow
tensorflow-graphics
scikit-learn
imageio
scipy
google-api-core
google-auth
google-auth-oauthlib
google-cloud-core
google-cloud-storage


openexr-numpy
